Flask/requirements.txt
absl-py==0.10.0
astor==0.8.1
backports.functools-lru-cache==1.6.1
backports.weakref==1.0.post1
bleach==1.5.0
cachetools==3.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
cycler==0.10.0
decorator==4.4.2
deskew==0.10.3
enum34==1.1.10
Flask==1.1.2
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.3.0
gast==0.2.2
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
html5lib==0.9999999
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.5
mock==3.0.5
networkx==2.2
numpy==1.16.6
oauthlib==3.1.0
opencv-python==3.1.0.0
opt-einsum==2.3.2
Pillow==6.2.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.0.3
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.5
scikit-image==0.14.5
scikit-learn==0.20.4
scipy==1.2.2
six==1.15.0
sklearn==0.0
subprocess32==3.5.4
tensorboard==1.6.0
tensorflow==1.6.0
tensorflow-estimator==2.0.1
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
